created config/constants.php instead of putting constants in .env, reverted change in older migrations and brought all changes in one place

+++ database/migrations/2026_02_13_164711_modify_fights_track_won.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Models\Track;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -11,25 +12,39 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasColumn('tracks', 'won')) {
+            Schema::table('tracks', function (Blueprint $table) {
+                $table->dropColumn('won');
+            });
+        }
         Schema::table('tracks', function (Blueprint $table) {
-        $table->dropColumn('won');
-    });
-    Schema::table('fights', function (Blueprint $table) {
-        $table->foreignId('won_track_id')->nullable()->constrained('tracks');
-    });
+            $table->unsignedTinyInteger('priority')->default(TRACK_PRIORITY_GUEST_ADDED)->change();
+            $table->boolean('used')->default(false)->change();
+        });
+        Schema::table('fights', function (Blueprint $table) {
+            $table->foreignId('won_track_id')->nullable()->constrained('tracks');
+        });
+
+        Track::query()->where('priority', 0)->update(['priority' => TRACK_PRIORITY_RESERVE]);
+        Track::query()->where('priority', 1)->update(['priority' => TRACK_PRIORITY_GUEST_ADDED]);
     }
 
     /**
      * Reverse the migrations.
      */
     public function down(): void
-    {   
-        
-        Schema::table('tracks', function (Blueprint $table) {
-        $table->boolean('won')->nullable();
-    });
-    Schema::table('fights', function (Blueprint $table) {
-        $table->dropForeign('won_track_id');
-    });
+    {
+        if (!Schema::hasColumn('tracks', 'won')) {
+            Schema::table('tracks', function (Blueprint $table) {
+                $table->boolean('won')->nullable();
+            });
+        }
+        if (Schema::hasColumn('fights', 'won_track_id')) {
+            Schema::table('fights', function (Blueprint $table) {
+                $table->dropForeign('fights_won_track_id_foreign');
+                $table->dropColumn('won_track_id');
+            });
+        }
+        Track::query()->update(['priority' => 0]);
     }
 };
